// frontend/src/api/bindings/Player.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Color } from "./Color";
import type { Faction } from "./Faction";
import type { Planet } from "./Planet";
import type { PlanetAttachment } from "./PlanetAttachment";
import type { Relic } from "./Relic";
import type { Technology } from "./Technology";

export type Player = { 
/**
 * The name of the player.
 */
name: string, 
/**
 * Which faction the player is playing.
 */
faction: Faction, 
/**
 * Which color the player has.
 */
color: Color, 
/**
 * Which planets the player controls and their attachments.
 */
planets: { [key in Planet]: Array<PlanetAttachment> }, 
/**
 * Which technologies the player has.
 */
technologies: Array<Technology>, 
/**
 * Which relics the player currently owns.
 */
relics: Array<Relic>, };